fixed updating other parts of a course like teachers, lessons and formulas from file

+++ web_for_msu_back/app/services/course_service.py
@@ -309,19 +309,23 @@ class CourseService:
             match key:
                 case "formulas":
                     formulas = course.formulas
-                    for i in range(min(len(formulas) - 1, len(data["formulas"]))):
+                    for i in range(min(len(formulas), len(data["formulas"]))):
+                        if formulas[i].name == "Баллы":
+                            continue
                         formulas[i].name = data["formulas"][i]["name"]
                         formulas[i].coefficient = data["formulas"][i]["coefficient"]
                     if len(data["formulas"]) < len(formulas) - 1:
-                        for i in range(len(data["formulas"]), len(formulas) - 1):
+                        for i in range(len(data["formulas"]), len(formulas)):
+                            if formulas[i].name == "Баллы":
+                                continue
                             self.db.session.delete(formulas[i])
                     elif len(data["formulas"]) > len(formulas) - 1:
-                        for i in range(len(formulas), len(data["formulas"])):
+                        for i in range(len(formulas) - 1, len(data["formulas"])):
                             course.formulas.append(Formula(course_id=course_id,
-                                                        name=data["formulas"][i]["name"],
-                                                        coefficient=data["formulas"][i]["coefficient"]))
+                                                           name=data["formulas"][i]["name"],
+                                                           coefficient=data["formulas"][i]["coefficient"]))
                 case "schedules":
-                    schedules = Schedule.query.where(Schedule.course_id==course_id).all()
+                    schedules = Schedule.query.where(Schedule.course_id == course_id).all()
                     for i in range(min(len(schedules), len(data["schedules"]))):
                         schedules[i].lesson_number = data["schedules"][i]["lesson_number"]
                         schedules[i].date = data["schedules"][i]["date"]
@@ -334,11 +338,11 @@ class CourseService:
                     elif len(data["schedules"]) > len(schedules):
                         for i in range(len(schedules), len(data["schedules"])):
                             course.lessons.append(Schedule(course_id=course_id,
-                                                         lesson_number=data["schedules"][i]["lesson_number"],
-                                                         date=data["schedules"][i]["date"],
-                                                         theme=data["schedules"][i]["theme"],
-                                                         plan=data["schedules"][i]["plan"],
-                                                         additional_info=data["schedules"][i]["additional_info"]))
+                                                           lesson_number=data["schedules"][i]["lesson_number"],
+                                                           date=data["schedules"][i]["date"],
+                                                           theme=data["schedules"][i]["theme"],
+                                                           plan=data["schedules"][i]["plan"],
+                                                           additional_info=data["schedules"][i]["additional_info"]))
                 case "teachers":
                     teachers = course.teachers
                     for teacher in teachers:
